requirements.txt
asgiref==3.7.2
boto3==1.29.0
botocore==1.32.0
dj-database-url==0.5.0
Django==3.2.23
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-environ==0.7.0
django-social-share==2.3.0
django-storages==1.14.2
django-summernote==0.8.20.0
gunicorn==21.2.0
jmespath==1.0.1
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
s3transfer==0.7.0
sqlparse==0.4.4
stripe==7.5.0
